parser: improve string truncating in fmt_string_from_mem

+++ include/core/trace.h
@@ -47,4 +47,4 @@ void cleanup_trace(struct trace_opts *opts, pid_t tracee);
  * @buf:    Buffer address (dest).
  * @size:   Bytes to read.
  */
-int read_tracee_mem(pid_t tracee, const void *addr, void *buf, size_t size);
+size_t read_tracee_mem(pid_t tracee, const void *addr, void *buf, size_t size);
